requirements.txt
-e .[all]
twine==5.1.1
mkdocs==1.6.0
pre-commit >=2.17.0,<4.0.0
pydantic==2.7.4
requests==2.32.3
pycryptodome==3.21.0
